redondeo de operaciones

+++ storage/src/sotrage_files.c
@@ -1,20 +1,7 @@
 #include "storage_files.h"
-#include "storage_bitmap.h"
-#include <sys/stat.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <errno.h>
-#include <commons/config.h>
-#include <commons/string.h>
-#include <commons/collections/list.h>
-#include <stdint.h>
-#include <unistd.h>
-#include <fcntl.h>
-
 
 // Crea un nuevo File:Tag en el FS
-int storage_create(char* file_name, char* tag)
+rta_storage storage_create(char* file_name, char* tag)
 {
     char* path = string_from_format("%s/files/%s", PUNTO_MONTAJE, file_name);
     mkdir(path, 0777);
@@ -23,8 +10,7 @@ int storage_create(char* file_name, char* tag)
     int saved_errno = errno;
     if (result_mkdir != 0 && saved_errno == EEXIST) {
         log_error(logger_storage, "Ya existe el file tag");
-        /* ERROR de preexistencia del file:tag*/
-        return STORAGE_RESULT_ERROR;
+        return ERR_PREEXISTENCIA;
     }
 
     char* logical_blocks_path = string_from_format("%s/%s", path, "logical_blocks");
@@ -38,14 +24,14 @@ int storage_create(char* file_name, char* tag)
     if (!meta)
     {
         log_error(logger_storage, "Error al crear metadata.config para %s:%s (%s)", file_name, tag, strerror(errno));
-        return STORAGE_RESULT_ERROR;
+        return NOT_DEF_ERR;
     }
 
     fprintf(meta, "TAMAÑO=0\nBLOCKS=[]\nESTADO=WORK_IN_PROGRESS\n");
     fclose(meta);
 
     log_info(logger_storage, "##CREATE - File:%s Tag:%s creado correctamente", file_name, tag);
-    return STORAGE_RESULT_OK;
+    return OP_EXITOSA;
 }
 
 
@@ -124,13 +110,13 @@ void storage_metadata_destroy(t_metadata_file* metadata)
     free(metadata);
 }
 
-int storage_truncate(const char* file_name, const char* tag, int new_size)
+rta_storage storage_truncate(const char* file_name, const char* tag, int new_size)
 {
     t_metadata_file* meta = storage_metadata_read(file_name, tag);
     if (!meta)
     {
         log_error(logger_storage, "No se pudo leer metadata de %s:%s para TRUNCATE", file_name, tag);
-        return STORAGE_RESULT_ERROR;
+        return NOT_DEF_ERR;
     }
 
     // Calcular bloques 
@@ -169,14 +155,23 @@ int storage_truncate(const char* file_name, const char* tag, int new_size)
             // Último bloque actual
             int last_index = list_size(meta->blocks) - 1;
             int block_to_free = atoi(list_get(meta->blocks, last_index));
+            char* path = string_from_format("%s/%05d.dat", logical_path, last_index);
 
             // Borrar el hardlink lógico
-            char* path = string_from_format("%s/%05d.dat", logical_path, last_index);
             unlink(path);
+            free(path);
 
             // Liberar del bitmap
             if (block_to_free != 0) {
-                mark_block_free(block_to_free);
+                struct stat blk_stat;
+                path = string_from_format("physical_blocks/block%04d.dat", block_to_free);
+                stat(path, &blk_stat);
+                if (blk_stat.st_nlink == 1)
+                {
+                    mark_block_free(block_to_free);
+                    unlink(path);
+                }
+                free(path);
             }
 
             // Quitar de metadata
@@ -190,16 +185,16 @@ int storage_truncate(const char* file_name, const char* tag, int new_size)
     storage_metadata_write(file_name, tag, meta);
 
     storage_metadata_destroy(meta);
-    return STORAGE_RESULT_OK;
+    return OP_EXITOSA;
 }
 
-int storage_write(char* file_name, char* tag, int l_block_num, char* buffer)
+rta_storage storage_write(char* file_name, char* tag, int l_block_num, char* buffer)
 {
     t_metadata_file* meta = storage_metadata_read(file_name, tag);
     if (!meta)
     {
         log_error(logger_storage, "No se pudo leer metadata para WRITE en %s:%s", file_name, tag);
-        return STORAGE_RESULT_ERROR;
+        return NOT_DEF_ERR;
     }
 
     // Valid que haya bloques asignados
@@ -207,55 +202,57 @@ int storage_write(char* file_name, char* tag, int l_block_num, char* buffer)
     {
         log_error(logger_storage, "WRITE falló: no hay bloques asignados en %s:%s", file_name, tag);
         storage_metadata_destroy(meta);
-        return STORAGE_RESULT_ERROR;
+        return ERR_FUERA_LIMITE;
     }
 
     int p_block_num = atoi(list_get(meta->blocks, l_block_num));
     
     char* phys_path = string_from_format("physical_blocks/block%04d.dat", p_block_num);
-    FILE* p_block = fopen(phys_path, "rb");
-    free(phys_path);
-    char* act_content = malloc(BLOCK_SIZE);
-    fread(act_content, BLOCK_SIZE, 1, p_block);
-    fclose(p_block);
-    // TODO: Check hash del nuevo y antiguo contenido, check links del bloque != 2 (si mismo y este archivo)
-
-    // Caso lindo: solo se reasigna el bloque fisico y el anterior sigue igual (por ej, el bloque fisico 0 que es para los bloques vacios)
-    char* log_path = string_from_format("files/%s/%s/logical_blocks/%05d.dat", file_name, tag, l_block_num);
-    unlink(log_path);
-    int new_p_block_num = find_free_block();
-    list_replace_and_destroy_element(meta->blocks, l_block_num, string_itoa(new_p_block_num), free);
-
-    phys_path = string_from_format("physical_blocks/block%04d.dat", new_p_block_num);
-    p_block = fopen(phys_path, "wb");
-    fwrite(buffer, BLOCK_SIZE, 1, p_block);
-    fclose(p_block);
-    link(phys_path, log_path);
-    free(phys_path);
-    free(log_path);
-
-    storage_metadata_write(file_name, tag, meta);
+    
+    struct stat blk_stat;
+    stat(phys_path, &blk_stat);
+    
+    if (blk_stat.st_nlink > 2 ) {
+        free(phys_path);
+        // Buscar nuevo bloque
+        p_block_num = find_free_block();
+        // Reemplazar en config
+        list_remove_and_destroy_element(meta->blocks, l_block_num, free);
+        list_add_in_index(meta->blocks, l_block_num, string_itoa(p_block_num));
+        storage_metadata_write(file_name, tag, meta);
+        // linkear nuevo bloque
+        char* log_path = string_from_format("files/%s/%s/logical_blocks/%05d.dat", file_name, tag, l_block_num);
+        unlink(log_path);
+        phys_path = string_from_format("physical_blocks/block%04d.dat", p_block_num);
+        link(phys_path, log_path);
+    }
+    
     storage_metadata_destroy(meta);
 
+    FILE* p_block = fopen(phys_path, "wb");
+    fwrite(buffer, BLOCK_SIZE, 1, p_block);
+    fclose(p_block);
+    free(phys_path);
+
     log_info(logger_storage, "WRITE completado en %s:%s (B = %d)", file_name, tag, l_block_num);
-    return STORAGE_RESULT_OK;
+    return OP_EXITOSA;
 }
 
 // el buffer ya debe tener un espacio de memoria asignado con un malloc(BLOCK_SIZE)
-int storage_read(char* file_name, char* tag, int l_block_num, char* buffer)
+rta_storage storage_read(char* file_name, char* tag, int l_block_num, char* buffer)
 {
     t_metadata_file* meta = storage_metadata_read(file_name, tag);
     if (!meta)
     {
         log_error(logger_storage, "No se pudo leer metadata para READ en %s:%s", file_name, tag);
-        return STORAGE_RESULT_ERROR;
+        return NOT_DEF_ERR;
     }
 
     if (l_block_num > list_size(meta->blocks))
     {
         log_warning(logger_storage, "READ fuera de rango en %s:%s (B = %d)", file_name, tag, l_block_num);
         storage_metadata_destroy(meta);
-        return STORAGE_RESULT_ERROR;
+        return ERR_FUERA_LIMITE;
     }
 
     int p_block_num = atoi(list_get(meta->blocks, l_block_num));
@@ -267,5 +264,106 @@ int storage_read(char* file_name, char* tag, int l_block_num, char* buffer)
     fclose(p_block);
 
     log_info(logger_storage, "READ completado en %s:%s (B = %d)", file_name, tag, l_block_num);
-    return STORAGE_RESULT_OK;
+    return OP_EXITOSA;
+}
+
+rta_storage storage_commit(char* file, char* tag) {
+    t_metadata_file* metadata = storage_metadata_read(file, tag);
+    free(metadata->estado);
+    metadata->estado = string_duplicate("COMMITED");
+    storage_metadata_write(file, tag, metadata);
+
+    // check hashes
+    int cant_blks = list_size(metadata->blocks);
+    for (int i = 0; i < cant_blks; i++) {
+        char* log_path = string_from_format("files/%s/%s/logical_blocks/%05d.dat", file, tag, i);
+        char* hash = get_hash_from_block(file, tag, i);
+        int actual_blk = atoi(list_get(metadata->blocks, i));
+        if (hash_is_loaded(hash)) {
+            char* phys_blk = block_asocied_to(hash);
+            char* phys_path = string_from_format("physical_blocks/%s.dat", phys_blk);
+            unlink(log_path);
+            link(phys_path, log_path);
+            free(phys_path);
+            mark_block_free(actual_blk);
+            list_remove_and_destroy_element(metadata->blocks, i, free);
+            list_add_in_index(metadata->blocks, i, string_itoa(atoi(phys_blk + 5))); // atoi -> itoa para tener un %d y no un %04d 
+            storage_metadata_write(file, tag, metadata);
+        } else {
+            load_hash_in_index(hash, actual_blk);
+        }
+    }
+    storage_metadata_destroy(metadata);
+    return OP_EXITOSA;
+}
+
+rta_storage storage_tag(char* file_o, char* tag_o, char* file_n, char* tag_n) {
+    rta_storage result = storage_create(file_n, tag_n);
+    if (result != OP_EXITOSA) {return result;}
+
+    t_metadata_file* metadata_ft_o = storage_metadata_read(file_o, tag_o);
+    free(metadata_ft_o->estado);
+    metadata_ft_o->estado = string_duplicate("WORK_IN_PROGRESS");
+    storage_metadata_write(file_n, tag_n, metadata_ft_o);
+
+    int cant_blks = list_size(metadata_ft_o->blocks);
+    for (int i = 0; i < cant_blks; i++) {
+        char* log_path = string_from_format("files/%s/%s/logical_blocks/%05d.dat", file_n, tag_n, i);
+        int blk_num = atoi(list_get(metadata_ft_o->blocks, i));
+        char* phys_path = string_from_format("physical_blocks/block%04d.dat", blk_num);
+        int r = link(phys_path, log_path);
+        free(phys_path);
+        free(log_path);
+        if (r != 0) { 
+            log_error(logger_storage, "no se pudo hacer el link() durante una instruccion TAG");
+            return NOT_DEF_ERR;
+        }
+    }
+
+    storage_metadata_destroy(metadata_ft_o);
+
+    return OP_EXITOSA;
+}
+
+rta_storage storage_delete(char* file, char* tag) {
+    t_metadata_file* meta = storage_metadata_read(file, tag);
+
+    int cant_blks = list_size(meta->blocks);
+    char* log_path = string_from_format("files/%s/%s", file, tag);
+    for (int i = 0; i < cant_blks; i++) {
+        int blk_num = atoi(list_get(meta->blocks, i));
+
+        char* hash = get_hash_from_block(file, tag, i);
+        remove_hash(hash);
+        free(hash);
+
+        char* log_block = string_from_format("%s/logical_blocks/%05d.dat", log_path, i);
+        char* phys_path = string_from_format("physical_blocks/block%04d.dat", blk_num);
+
+        unlink(log_block);
+        free(log_block);
+
+        struct stat blk_stat;
+        stat(phys_path, &blk_stat);
+        
+        if (blk_stat.st_nlink < 2) {
+            unlink(phys_path);
+            mark_block_free(blk_num);
+        }
+        free(phys_path);
+    }
+    storage_metadata_destroy(meta);
+
+    char* meta_conf = string_from_format("%s/metadata.config", log_path);
+    remove(meta_conf);
+    free(meta_conf);
+
+    char* log_blks = string_from_format("%s/logical_blocks", log_path);
+    rmdir(log_blks);
+    free(log_blks);
+    
+    rmdir(log_path);
+    free(log_path);
+
+    return OP_EXITOSA;
 }
